# swagger.yaml
openapi: 3.0.3
info:
  title: SmartTransit SMS Authentication API
  description: |
    SMS-based authentication backend for SmartTransit mobile applications.

    ## Features
    - SMS OTP authentication for passengers
    - Staff registration and verification
    - JWT token-based authorization
    - Profile management
    - Rate limiting and security
    - Automated trip generation from timetables/schedules

    ## Authentication Flow
    1. Request OTP via `/auth/send-otp`
    2. Verify OTP via `/auth/verify-otp` or `/auth/verify-otp-staff`
    3. Receive JWT access & refresh tokens
    4. Use access token for protected endpoints
    5. Refresh token when access token expires

    ## Staff Employment Model
    Staff members (drivers/conductors) have a separated profile and employment structure:
    - **Staff Profile (`bus_staff`)**: Personal information, license details, verification status
    - **Employment Records (`bus_staff_employment`)**: Employment history with bus owners

    **Employment Lifecycle:**
    1. Staff registers profile (no bus owner link initially)
    2. Bus owner verifies staff via phone number
    3. Bus owner links staff (creates employment record)
    4. Employment can be terminated (record preserved for history)
    5. Staff can be linked to new bus owner (new employment record)

    **Benefits:**
    - Complete employment history tracking
    - Staff can move between companies
    - Performance metrics per employment period
...

...

...

...

...

...
  contact:
...
servers:
...
paths:
...
  /api/v1/staff/check-registration:
    post:
      summary: Check staff registration status
      description: Check if a phone number is registered as staff and get registration details
      operationId: checkStaffRegistration
      tags:
        - Staff
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  pattern: '^\+94[0-9]{9}$'
                  example: "+94771234567"
      responses:
        "200":
          description: Registration status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_registered:
                    type: boolean
                    example: true
                  staff:
                    $ref: "#/components/schemas/BusStaff"
        "404":
          description: Phone number not registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_registered:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Phone number not registered as staff
  /api/v1/staff/register:
    post:
      summary: Register new staff member
      description: |
        Register a new driver or conductor profile.

        **New Registration Flow:**
        - Staff registers their profile (driver or conductor)
        - Profile is created without any bus owner link
        - Staff waits for a bus owner to link them to their organization
        - Employment details (hire date, status, performance) are managed separately

        **Requirements:**
        - Valid phone number (not already registered as staff)
        - Staff type (driver or conductor)
        - Driver-specific: license number and expiry date
      operationId: registerStaff
      tags:
        - Staff
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - last_name
                - staff_type
              properties:
                first_name:
                  type: string
                  description: Staff member's first name
                  example: "Kamal"
                last_name:
                  type: string
                  description: Staff member's last name
                  example: "Perera"
                staff_type:
                  type: string
                  enum: [driver, conductor]
                  description: Type of staff - driver or conductor
                  example: "driver"
                license_number:
                  type: string
                  description: NTC license number (required for drivers)
                  example: "NTC1234567"
                license_expiry_date:
                  type: string
                  format: date
                  description: License expiry date (required for drivers)
                  example: "2026-12-31"
                experience_years:
                  type: integer
                  description: Years of experience
                  example: 5
                emergency_contact:
                  type: string
                  description: Emergency contact phone number
                  example: "+94777654321"
                emergency_contact_name:
                  type: string
                  description: Emergency contact person name
                  example: "Nimal Perera"
      responses:
        "201":
          description: Staff registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Staff registered successfully
                  staff:
                    $ref: "#/components/schemas/BusStaff"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/staff/bus-owners/search:
    get:
      summary: Search bus owners
      description: Search for bus owners by name or NIC
      operationId: searchBusOwners
      tags:
        - Staff
      security: []
      parameters:
        - name: query
          in: query
          required: true
          description: Search query (name or NIC)
          schema:
            type: string
            example: "Perera"
        - name: limit
          in: query
          required: false
          description: Maximum number of results
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  bus_owners:
                    type: array
                    items:
                      $ref: "#/components/schemas/BusOwner"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/staff/profile:
    get:
      summary: Get staff profile
      description: Retrieve the authenticated staff member's profile
      operationId: getStaffProfile
      tags:
        - Staff
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Staff profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: Update staff profile
      description: Update the authenticated staff member's profile
      operationId: updateStaffProfile
      tags:
        - Staff
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Driver"
                nic:
                  type: string
                  example: "199512345678"
                license_number:
                  type: string
                  example: "B1234567"
                address:
                  type: string
                  example: "123 Main St, Colombo"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
                  staff:
                    $ref: "#/components/schemas/BusStaff"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================================
  # Bus Owner Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/auth/verify-otp` or `/auth/verify-otp-staff`.

        Include in Authorization header as: `Bearer <token>`
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        phone:
          type: string
          example: "+94771234567"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        nic:
          type: string
          example: "199512345678"
        address:
          type: string
          example: "123 Main St, Colombo"
        role:
          type: string
          enum: [passenger, driver, conductor]
          example: passenger
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-10-19T07:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-10-19T07:00:00Z"
...
    BusStaff:
      type: object
      description: |
        Staff member profile (driver or conductor).
        Contains profile information only - employment details are in BusStaffEmployment.
      properties:
        id:
          type: string
          format: uuid
          example: "d581c281-ba77-49d2-8f68-5f7377417bf8"
        user_id:
          type: string
          format: uuid
          example: "3edeb9af-51c3-4c24-a5d0-76c9dee2d4ea"
        first_name:
          type: string
          example: "Kamal"
        last_name:
          type: string
          example: "Perera"
        staff_type:
          type: string
          enum: [driver, conductor]
          example: "driver"
        license_number:
          type: string
          example: "NTC1234567"
        license_expiry_date:
          type: string
          format: date
          example: "2026-12-31"
        experience_years:
          type: integer
          example: 5
        emergency_contact:
          type: string
          example: "+94777654321"
        emergency_contact_name:
          type: string
          example: "Nimal Perera"
        medical_certificate_expiry:
          type: string
          format: date
          nullable: true
          example: "2025-06-30"
        background_check_status:
          type: string
          enum: [pending, in_progress, approved, rejected]
          example: "pending"
        profile_completed:
          type: boolean
          example: true
        verification_notes:
          type: string
          nullable: true
          example: "Documents verified"
        verified_at:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-15T10:30:00Z"
        verified_by:
          type: string
          format: uuid
          nullable: true
          description: Admin user ID who verified this staff
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BusStaffEmployment:
      type: object
      description: |
        Employment record for a staff member.
        Tracks employment history when staff moves between bus owners.
      properties:
        id:
          type: string
          format: uuid
          example: "e692d392-cb88-50e3-9f79-6f8488528cf9"
        staff_id:
          type: string
          format: uuid
          example: "d581c281-ba77-49d2-8f68-5f7377417bf8"
        bus_owner_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        employment_status:
          type: string
          enum: [active, inactive, suspended, terminated]
          example: "active"
        hire_date:
          type: string
          format: date
          example: "2024-01-15"
        termination_date:
          type: string
          format: date
          nullable: true
          example: null
        termination_reason:
          type: string
          nullable: true
          example: null
        salary_amount:
          type: number
          format: decimal
          nullable: true
          example: 75000.00
        performance_rating:
          type: number
          format: double
          nullable: true
          example: 4.5
        total_trips_completed:
          type: integer
          example: 150
        is_current:
          type: boolean
          description: Whether this is the current active employment
          example: true
        notes:
          type: string
          nullable: true
          example: "Excellent performance"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BusOwner:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 2
        company_name:
          type: string
          example: "Express Transport Ltd"
        nic:
          type: string
          example: "197512345678"
        created_at:
          type: string
          format: date-time
...
    StaffProfile:
      type: object
      description: |
        Complete staff profile including user info, staff details, and current employment.
      properties:
        user:
          $ref: "#/components/schemas/User"
        staff:
          $ref: "#/components/schemas/BusStaff"
        employment:
          $ref: "#/components/schemas/BusStaffEmployment"
          nullable: true
          description: Current employment record (null if not linked to any bus owner)
        bus_owner:
          $ref: "#/components/schemas/BusOwner"
          nullable: true
          description: Current employer's bus owner profile (null if not employed)
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        details:
          type: string
          example: Phone number is required
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalidPhone:
              value:
                error: Invalid phone number
                details: Phone number must be in E.164 format (+94XXXXXXXXX)
            missingField:
              value:
                error: Missing required field
                details: Field 'phone' is required
    Unauthorized:
      description: Unauthorized - Invalid or expired token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalidToken:
              value:
                error: Unauthorized
                details: Invalid or expired token
            missingToken:
              value:
                error: Unauthorized
                details: Authorization header required
...
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            phoneExists:
              value:
                error: Phone number already registered
                details: This phone number is already in use
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            serverError:
              value:
                error: Internal server error
                details: An unexpected error occurred